requirements.txt
click==6.6
coverage==4.2
eventlet==0.19.0
Flask==0.11.1
funcsigs==1.0.2
greenlet==0.4.10
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
mock==2.0.0
pbr==1.10.0
py==1.4.31
pyserial==3.1.1
pytest==3.0.2
pytest-cov==2.3.1
requests==2.11.1
six==1.10.0
SQLAlchemy==1.0.15
Werkzeug==0.11.11
wheel==0.24.0
